fix(Tickets): 调整工单详情样式
