+++ src/views/tickets/RequestAssetPerm/Detail/TicketDetail.vue
@@ -1,147 +1,38 @@
 <template>
-  <el-row>
-    <el-col :span="17">
-      <IBox class="box">
-        <div slot="header" class="clearfix ibox-title">
-          <i class="fa fa-info-circle" /> {{ $t('common.BasicInfo') }}
-        </div>
-        <div class="content">
-          <el-row :gutter="10">
-            <el-col v-for="item in detailCardItems" :key="'card-' + item.key" :span="12">
-              <el-row class="item">
-                <el-col :span="6">
-                  <!--        </div>-->
-                  <div :style="{ 'text-align': 'align' }" class="item-label">
-                    <label>{{ item.key }}: </label>
-                  </div>
-                </el-col>
-                <el-col :span="18">
-                  <div class="item-text">
-                    <ItemValue :value="item.value" v-bind="item" />
-                  </div>
-                </el-col>
-              </el-row>
-            </el-col>
-          </el-row>
-          <el-divider />
-          <el-row :gutter="10">
-            <el-col v-for="item in spcialCardItems" :key="'card-' + item.key" :span="12">
-              <el-row class="item">
-                <el-col :span="6">
-                  <!--        </div>-->
-                  <div :style="{ 'text-align': 'align' }" class="item-label">
-                    <label>{{ item.key }}: </label>
-                  </div>
-                </el-col>
-                <el-col :span="18">
-                  <div class="item-text">
-                    <ItemValue :value="item.value" v-bind="item" />
-                  </div>
-                </el-col>
-              </el-row>
-            </el-col>
-          </el-row>
-        </div>
-
-      </IBox>
-      <IBox v-if="hasActionPerm&&object.status !== 'closed'" class="box">
-        <div slot="header" class="clearfix ibox-title">
-          <i class="fa fa-info-circle" /> {{ $t('common.Actions') }}
-        </div>
-        <template>
-          <el-form ref="request_form" :model="request_form" label-width="140px" label-position="left" class="assets">
-            <el-form-item :label="$t('tickets.Asset')" required>
-              <Select2 ref="select2" v-model="request_form.asset" v-bind="asset_select2" style="width: 30% !important" />
-            </el-form-item>
-            <el-form-item :label="$t('tickets.SystemUser')" required>
-              <Select2 ref="select2" v-model="request_form.systemuser" v-bind="systemuser_select2" style="width: 30% !important" />
-            </el-form-item>
-          </el-form>
-        </template>
-      </IBox>
-      <IBox class="box">
-        <div slot="header" class="clearfix ibox-title">
-          <i class="fa fa-info-circle" /> {{ $t('common.Comment') }}
-        </div>
-        <div class="feed-activity-list">
-          <div class="feed-element">
-            <a href="#" class="pull-left">
-              <el-avatar :src="imageUrl" class="header-avatar" />
-            </a>
-            <div class="media-body ">
-              <strong>{{ object.user_display }}</strong> <small class="text-muted"> {{ formatTime(object.date_created) }}</small>
-              <br>
-              <small class="text-muted">{{ toSafeLocalDateStr(object.date_created) }} </small>
-              <div style="padding-top: 10px">
-                <span v-html="object.body" />
-              </div>
-            </div>
-          </div>
-        </div>
-        <template v-if="comments">
-          <div v-for="item in comments" :key="item.user_display + item.body">
-            <div class="feed-element">
-              <a href="#" class="pull-left">
-                <el-avatar :src="imageUrl" class="header-avatar" />
-              </a>
-              <div class="media-body ">
-                <strong>{{ item.user_display }}</strong> <small class="text-muted">{{ formatTime(item.date_created) }}</small>
-                <br>
-                <small class="text-muted">{{ toSafeLocalDateStr(item.date_created) }}</small>
-                <div style="padding-top: 10px">
-                  {{ item.body }}
-                </div>
-              </div>
-            </div>
-          </div>
-        </template>
-        <el-form ref="comments" :model="form" label-width="45px" style="padding-top: 20px">
-          <el-form-item :label="$t('tickets.reply')">
-            <el-input v-model="form.comments" :autosize="{ minRows: 4 }" type="textarea" />
+  <GenericTicketDetail
+    :object="object"
+    :detail-card-items="detailCardItems"
+    :special-card-items="specialCardItems"
+    :approve="handleApprove"
+  >
+    <IBox v-if="hasActionPerm&&object.status !== 'closed'" class="box">
+      <div slot="header" class="clearfix ibox-title">
+        <i class="fa fa-info-circle" /> {{ $t('common.Actions') }}
+      </div>
+      <template>
+        <el-form ref="request_form" :model="request_form" label-width="140px" label-position="left" class="assets">
+          <el-form-item :label="$t('tickets.Asset')" required>
+            <Select2 ref="select2" v-model="request_form.asset" v-bind="asset_select2" style="width: 30% !important" />
           </el-form-item>
-          <el-form-item style="float: right">
-            <template v-if="hasActionPerm">
-              <el-button :disabled="object.status === 'closed'" type="primary" size="small" @click="handleApprove"><i class="fa fa-check" />{{ $t('tickets.Accept') }}</el-button>
-              <el-button :disabled="object.status === 'closed'" type="warning" size="small" @click="handleReject"><i class="fa fa-ban" />{{ $t('tickets.Reject') }}</el-button>
-            </template>
-            <el-button :disabled="object.status === 'closed'" type="danger" size="small" @click="handleClosed"><i class="fa fa-times" />{{ $t('tickets.Close') }}</el-button>
-            <el-button :disabled="object.status === 'closed'" type="info" size="small" @click="handleComment"><i class="fa fa-pencil" />{{ $t('tickets.Comment') }}</el-button>
+          <el-form-item :label="$t('tickets.SystemUser')" required>
+            <Select2 ref="select2" v-model="request_form.systemuser" v-bind="systemuser_select2" style="width: 30% !important" />
           </el-form-item>
         </el-form>
-      </IBox>
-    </el-col>
-    <el-col :span="6" :offset="1">
-      <IBox>
-        <div style="height: 350px;">
-          <el-steps direction="vertical" :active="ticketStatus">
-            <el-step
-              :title="`${this.$t('tickets.OpenTicket')}：${object.type_display}`"
-              :description="`${this.$t('tickets.Applicant')}：${object.user_display}`"
-            />
-            <el-step
-              :title="`${this.$t('tickets.HandleTicket')}`"
-              :description="`${this.$t('tickets.Assignees')}：${object.assignees_display}`"
-            />
-            <el-step
-              :title="`${this.$t('tickets.FinishedTicket')}`"
-              :description="`${this.$t('tickets.Assignee')}：${object.assignee_display}`"
-            />
-          </el-steps>
-        </div>
-      </IBox>
-    </el-col>
-  </el-row>
+      </template>
+    </IBox>
+  </GenericTicketDetail>
 </template>
 
 <script>
-import IBox from '@/components/IBox'
-import ItemValue from './ItemValue'
-import Select2 from '@/components/Select2'
+import { formatTime, getDateTimeStamp } from '@/utils/index'
 import { toSafeLocalDateStr } from '@/utils/common'
-import { formatTime, getDateTimeStamp } from '@/utils'
+import { STATUS_MAP } from '../../const'
+import Select2 from '@/components/Select2'
+import IBox from '@/components/IBox'
+import GenericTicketDetail from '@/views/tickets/components/GenericTicketDetail'
 export default {
   name: '',
-  components: { IBox, ItemValue, Select2 },
+  components: { GenericTicketDetail, IBox, Select2 },
   props: {
     object: {
       type: Object,
@@ -150,11 +41,7 @@ export default {
   },
   data() {
     return {
-      imageUrl: require('@/assets/img/admin.png'),
-      form: {
-        comments: '',
-        confirmed_assets: []
-      },
+      statusMap: this.object.status === 'open' ? STATUS_MAP[this.object.status] : STATUS_MAP[this.object.action],
       request_form: {
         asset: this.object.confirmed_assets,
         systemuser: ''
@@ -184,18 +71,12 @@ export default {
   },
   computed: {
     detailCardItems() {
-      const vm = this
       return [
         {
           key: this.$t('tickets.status'),
           value: this.object.status,
-          formatter: function(row, data) {
-            const open = vm.$t('tickets.Open')
-            const close = vm.$t('tickets.Close')
-            if (data === 'open') {
-              return <el-tag type='primary' size='mini'>{open}</el-tag>
-            }
-            return <el-tag type='danger' size='mini'>{close}</el-tag>
+          formatter: (item, val) => {
+            return <el-tag type={this.statusMap.type} size='mini'> { this.statusMap.title }</el-tag>
           }
         },
         {
@@ -220,7 +101,7 @@ export default {
         }
       ]
     },
-    spcialCardItems() {
+    specialCardItems() {
       return [
         // {
         //   key: this.$t('tickets.Assignee'),
@@ -246,18 +127,7 @@ export default {
     },
     hasActionPerm() {
       return this.object.assignees.indexOf(this.$store.state.users.profile.id) !== -1
-    },
-    ticketStatus() {
-      return this.object.status === 'open' ? 2 : 3
     }
-  },
-  mounted() {
-    const url = `/api/v1/tickets/tickets/${this.object.id}/comments/`
-    this.$axios.get(url).then(res => {
-      this.comments = res
-    }).catch(err => {
-      this.$message.error(err)
-    })
   },
   methods: {
     formatTime(dateStr) {
@@ -287,7 +157,6 @@ export default {
       })
     },
     handleApprove() {
-      // this.$axios.patch(url, data).then(res => this.reloadPage()).catch(err => this.$message.error(err))
       if (this.request_form.asset.length === 0 || this.request_form.systemuser === '') {
         return this.$message.error(this.$t('common.NeedAssetsAndSystemUserErrMsg'))
       } else {
@@ -305,50 +174,17 @@ export default {
           () => this.$message.success(this.$t('common.updateErrorMsg'))
         )
       }
-    },
-    handleReject() {
-      this.createComment(function() {})
-      const url = `/api/v1/tickets/tickets/${this.object.id}/`
-      const data = { action: 'reject' }
-      this.$axios.patch(url, data).then(res => this.reloadPage()).catch(err => this.$message.error(err))
-    },
-    handleClosed() {
-      const url = `/api/v1/tickets/tickets/${this.object.id}/`
-      const data = { status: 'closed' }
-      this.$axios.patch(url, data).then(res => this.reloadPage()).catch(err => this.$message.error(err))
-    },
-    handleComment() {
-      this.createComment()
     }
   }
 }
 </script>
 
 <style scoped>
-  .content {
-    font-size: 13px;
-    line-height: 2.5;
-  }
   .assets{
     margin-top: 14px;
   }
-  .feed-activity-list {
-    padding-top: 20px;
-    line-height: 1.5;
-  }
   .feed-activity-list .feed-element {
     border-bottom: 1px solid #e7eaec;
-  }
-  .feed-element:first-child {
-    margin-top: 0;
-  }
-  .feed-element {
-    padding-top: 15px;
-    padding-bottom: 15px;
-  }
-  .feed-element,
-  .media-body {
-    overflow: hidden;
   }
   .feed-element > .pull-left {
     margin-right: 10px;
@@ -356,12 +192,6 @@ export default {
   .feed-element .header-avatar {
     width: 38px;
     height: 38px;
-  }
-  .text-muted {
-    color: #888888;
-  }
-  .el-button--mini {
-    padding: 4px 6px;
   }
   .box {
     margin-bottom: 15px;
